Remove base fusion
